feat: enhance layout with navigation and footer, add custom fonts and active section hook
