+++ src/app/globals.css
@@ -1,4 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Sonsie+One&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Jersey+20&display=swap');
+@import url('https://fonts.cdnfonts.com/css/geist-mono');
 @import 'tailwindcss';
 
 :root {
@@ -21,6 +23,21 @@ body {
 
 .font-sonsie {
   font-family: 'Sonsie One', cursive;
+}
+
+.font-inter-heavy {
+  font-family: 'Inter', sans-serif;
+  font-weight: 800;
+}
+
+.font-jersey {
+  font-family: 'Jersey 20', cursive;
+  font-size: 75px;
+}
+
+.font-geist-mono {
+  font-family: 'Geist Mono', monospace;
+  font-weight: 700;
 }
 
 .title {
