+++ src/app/layout.tsx
@@ -1,5 +1,7 @@
 import type { Metadata } from 'next';
 import './globals.css';
+import Image from 'next/image';
+import Navigation from '@/components/Navigation';
 
 export const metadata: Metadata = {
   title: 'Cristian Sîrbu',
@@ -14,7 +16,21 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body>
-        {children}
+        <Navigation />
+        <main>{children}</main>
+        <footer className="w-full -mt-8 mb-4">
+          <p className="w-full text-center font-medium text-xs text-[#86868B]">
+            Designed <code>&&</code> Built with
+            <Image
+              src="/icons/misc/heart.svg"
+              alt="heart"
+              width={12}
+              height={12}
+              className="inline-block mx-1"
+            />
+            by Cristian Sîrbu
+          </p>
+        </footer>
       </body>
     </html>
   );
